+++ Molten.Math/Vectors/SByte2.v2.cs
@@ -21,7 +21,7 @@ namespace Molten
         /// <param name="right">Second source vector.</param>
         public static sbyte Cross(ref SByte2 left, ref SByte2 right)
         {
-            return (left.X * right.Y) - (left.Y * right.X);
+            return (sbyte)((left.X * right.Y) - (left.Y * right.X));
         }
 
         /// <summary>
@@ -31,7 +31,7 @@ namespace Molten
         /// <param name="right">Second source vector.</param>
         public static sbyte Cross(SByte2 left, SByte2 right)
         {
-            return (left.X * right.Y) - (left.Y * right.X);
+            return (sbyte)((left.X * right.Y) - (left.Y * right.X));
         }
 #endregion
 	}
